@@ -1,12 +1,7 @@
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Microsoft.AspNetCore.Blazor.Hosting;
-using System.ComponentModel;
+﻿using Microsoft.AspNetCore.Blazor.Hosting;
 
 namespace MLS.Blazor
 {
-    [EditorBrowsable(EditorBrowsableState.Never)]
     public class Program
     {
         public static void Main(string[] args)
